requirements.txt
h5py==3.12.0
huggingface_hub==0.22.2
hydra-core==1.3.2
matplotlib
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.9.0.80
Pillow==11.3.0
ptlflow==0.4.1
scenedetect==0.6.5.2
scikit_learn==1.6.1
scipy==1.13.1
scikit-image
timm==1.0.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
wandb==0.16.6
